test_project/requirements.txt
asgiref==3.9.1 ; python_version >= "3.12"
colorama==0.4.6 ; python_version >= "3.12" and sys_platform == "win32"
django-filter==25.1 ; python_version >= "3.12"
django==5.2.4 ; python_version >= "3.12"
djangorestframework==3.16.0 ; python_version >= "3.12"
iniconfig==2.1.0 ; python_version >= "3.12"
packaging==25.0 ; python_version >= "3.12"
pluggy==1.6.0 ; python_version >= "3.12"
pygments==2.19.2 ; python_version >= "3.12"
pytest-django==4.11.1 ; python_version >= "3.12"
pytest==8.4.1 ; python_version >= "3.12"
sqlparse==0.5.3 ; python_version >= "3.12"
tzdata==2025.2 ; python_version >= "3.12" and sys_platform == "win32"
